+++ clients/ios/Classes/AddSiteViewController.m
@@ -185,7 +185,7 @@
         return;
     }
     
-    int periodLoc = [phrase rangeOfString:@"."].location;
+    NSInteger periodLoc = [phrase rangeOfString:@"."].location;
     if (periodLoc != NSNotFound && siteAddressInput.returnKeyType != UIReturnKeyDone) {
         // URL
         [siteAddressInput setReturnKeyType:UIReturnKeyDone];
@@ -302,11 +302,11 @@
 
 - (NSString *)extractParentFolder {
     NSString *parent_folder = [inFolderInput text];
-    int folder_loc = [parent_folder rangeOfString:@" - " options:NSBackwardsSearch].location;
+    NSInteger folder_loc = [parent_folder rangeOfString:@" - " options:NSBackwardsSearch].location;
     if ([parent_folder length] && folder_loc != NSNotFound) {
         parent_folder = [parent_folder substringFromIndex:(folder_loc + 3)];
     }
-    int top_level_loc = [parent_folder rangeOfString:@" Top Level " options:NSBackwardsSearch].location;
+    NSInteger top_level_loc = [parent_folder rangeOfString:@" Top Level " options:NSBackwardsSearch].location;
     if (parent_folder.length && top_level_loc != NSNotFound) {
         parent_folder = @"";
     }
@@ -425,7 +425,7 @@ numberOfRowsInComponent:(NSInteger)component {
 #pragma mark Autocomplete sites
 
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return [autocompleteResults count];
 }
